imporve MSLi recontruction after AX typings

+++ include/msl/mslgcn.h
@@ -16,7 +16,7 @@ void mslCreateLogTable(void);
 void MSL_ClearVolatileFlag(unsigned long request_address);
 
 extern int g_bMSL_GCN_BREAK;
-extern int g_MSL_volatile_flag;
+extern unsigned long g_MSL_volatile_flag;
 extern unsigned long mslGCN_AXCallback_Ticks;
 extern int debugger_mbo1;
 extern int debugger_at1;
